enhance: move Missing Data Section to data tab
